Extract EventsList, write todos
